use checksum to check if there is a new update
When restarting the Docker container often, it is easy to hit the limit
for a given token. Use the checksum API endpoint to check if we have an
update available.

The parsing is a bit crude to avoid pulling jq. The error handling does
not need to be very robust as if the compute the wrong checksum, it
won't match the local one and we will trigger a download.